pyproject.toml
[project]
name = "causal-models-ai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "causal-learn",
    "numpy",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "pydot",
    "ipython",
    "ipykernel",
    "jupyter",
    "anthropic",
    "openai",
    "google-generativeai",
    "jinja2",
    "pyyaml",
    "pydantic",
]
